refactor: Update button styles and colors
- Update buttonStyle.css to change the background color and hover color for customFab and timerFab classes
- Update projectOverviewStyle.css to change the background color for the session class
- Update styles.css to change the background color and header color

Refactor the button styles and colors to improve the visual appearance of the application. The background color and hover color for the customFab and timerFab classes in buttonStyle.css are updated to #365E32 and #81A263 respectively. The background color for the session class in projectOverviewStyle.css is updated to #FD9B63. The background color and header color in styles.css are updated to #E7D37F and #ffffff respectively.